Divide main.py into some small .py file, change audio file, change the difficult of game
