+++ audio.py
@@ -1,5 +1,6 @@
 import pygame as py
 import random 
+from ui.utils import *
 
 folder = "audio/"
 
@@ -30,11 +31,12 @@ zombie_appear_dict = ["digging_0", "digging_1"]
 py.mixer.init()
 py.mixer.set_num_channels(32)
 
-class sfx:
+class audio:
     def __init__(self):
         self.zombie_groan_sound = [py.mixer.Sound(mp3(x)) for x in zombie_groan_dict]
         self.zombie_appear_sound = [py.mixer.Sound(mp3(x)) for x in zombie_appear_dict]
         self.sounds = [py.mixer.Sound(mp3(x)) for x in music_dict]
+        self.sounds_0 = audio_importer('assets', 'audio')
 
     # Sound.play(loops, max time, fade-in)
     # general play sound
@@ -58,13 +60,24 @@ class sfx:
     def play_start_sound(self):
         self.play(self.sounds[START], 1, 0)
 
+    def play_awooga_sound(self):
+        self.play(self.sounds_0['awooga'], 1, 0)
+
     # play when game started in first 1 minute
+    def play_grasswalk(self):
+        self.play(self.sounds_0['grasswalk'], 1, -1, 0, 1000)
+
+    def stop_grasswalk(self):
+        self.stop(self.sounds_0['grasswalk'])
+
+    # play when game started in next 1 minute
     def play_looboon(self):
         self.play(self.sounds[LOOBOON], 1, -1, 0, 1000)
 
     def stop_looboon(self):
         self.stop(self.sounds[LOOBOON])
 
+    # play when game started after 2 minutes
     def play_brain_maniac(self):
         self.play(self.sounds[BRAINIAC_MANIAC], 1, -1, 0, 1000)
 
@@ -75,6 +88,12 @@ class sfx:
     def play_lose_sound(self):
         self.play(self.sounds[LOSE], 1, 0)
 
+    def play_losemusic_sound(self):
+        self.play(self.sounds_0['losemusic'], 1, 0)
+
+    def play_scream_sound(self):
+        self.play(self.sounds_0['scream'], 1, 0)
+
     # play when zombie come our house
     def play_eat_sound(self):
         self.play(self.sounds[EAT], 1, 0)
@@ -82,6 +101,9 @@ class sfx:
     # play when hit zombie
     def play_bonk_sound(self):
         self.play(self.sounds[BONK], 1, 0)
+
+    def play_cherrybomb(self):
+        self.play(self.sounds_0['cherrybomb'], 1, 0)
 
     # every 3 seconds check if there is a zombie in screen then play this sound
     def play_zombie_groan(self):
@@ -91,6 +113,6 @@ class sfx:
     # play when zombie appear
     def play_zombie_appear(self):
         rd = random.randint(0,1)
-        self.play(self.zombie_appear_sound[rd], 0.3)
+        self.play(self.zombie_appear_sound[rd], 0.2)
