--- ui/settings.py
@@ -1,3 +1,0 @@
-import pygame
-from os import walk
-from os.path import join
